DateString added to helpers folder
